requirements.txt
Django
psycopg2
wheel
python-magic
djangorestframework
pyyaml
uritemplate
selenium
coreapi
django-filter
drf-yasg